requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
infobip-api-python-sdk==5.0.1
install==1.3.5
packaging==23.2
phonenumbers==8.13.17
Pillow==10.1.0
pycountry==22.3.5
pydantic==1.10.13
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2022.3.15
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
